Fixed LFTR bugs
LFTR now applicable to all strength reduction candidates with 2 variable operands.
If LFTR forms an expression that has already been processed by EPRE, it will not be processed again.
Implemented FindScalarVersion() to determine the correct SSA version to use when LFTR causes a var to appear at the compare.
The isRebuild parameter for BuildWorkListExpr is now split into sRebuild and insertSorted for its 2 different purposes.